Bug: DelayedLoadRelations uses inconsistent quoting for string-literals. (#95)
* Bug: DelayedLoadRelations uses inconsistent quoting for string-literals.

If using HQL on an Entity that is implicitly joined to other Entities via @OneToOne, @OneToMany, @ManyToOne, or @ManyToMany,
and an ID value has a single-quote (') in it, the HQL-escaping ('') is not unescaped, and it gets escaped into the dialect,
resulting in (\\'\\'), and a query that fails to join.

Example debug statements:
===
[trace] session.d:1207:delayedLoadRelations unknownKeys = [Bucky O'Hare]
[trace] session.d:1208:delayedLoadRelations createCommaSeparatedKeyList(unknownKeys) = 'Bucky O''Hare'
                                                               extra quote is being added here ^
[trace] session.d:1290:listObjects SQL: SELECT _t1.contract_code, ... FROM finance_contract_refs AS _t1 WHERE _t1.contract_code IN ( 'Bucky O\\'\\'Hare')
===

The fix is to modify the HQL Parser to unescape ('') as ('), and then let specific dialects re-escape the ('). PostgreSQL, by default,
does not support C-style escape characters unless quoted as E'hello\nthere\n' (note the leading (E)). However, all dialects support
the SQL standard ('') syntax.

* Remove debug output.

* Add another quoting test with the string literal directly in the query rather than a parameter.